requirements.txt
# General dependencies
python-dotenv
ipykernel

# AWS Strands
strands-agents
strands-agents-tools